requirements.txt
pyyaml
requests
beautifulsoup4